Add support for using a custom error handler
The default behaviour is to throw when ever we fail to read or write the
config. This is fine for most cases, but some might want to customize
that. For example you can send an error log to your server instead of
throwing. Or maybe you are storing non-essential data so you just don't
want to throw.